LPS-115716 Move classes to *.constants
