+++ modules/dxp/apps/vldap/vldap-server/src/main/java/com/liferay/vldap/server/internal/constants/OIDConstants.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.vldap.server.internal.util;
+package com.liferay.vldap.server.internal.constants;
 
 /**
  * @author Jonathan Potter
